Add an example to show RepeatEvent usage
Notice this les example exhibit a bug in the simulator: when
several events are scheduled at the same time, the order is not
(and should not be) deterministic and the result is unpredictible.

The right thing to do should be to base every action on the same
date - not to update data after /each/ event.